test-requirements.txt
# Requirements to test scd


bandit
fauxfactory
flake8
flake8-builtins
flake8-coding
flake8_comprehensions
flake8-debugger
flake8-import-order
flake8-mutable
flake8-quotes
flake8-string-format
pep8-naming
pycodestyle
pytest
pytest-cov
pytest-fauxfactory
pytest-profiling
PyYAML
radon
toml
xenon
